Core: Moved database/user to a views directory
