+++ internal/database/database.go
@@ -1,5 +1,7 @@
 package database
 
+import "github.com/Serdok/pokemon-go/internal/database/views"
+
 type Storage struct {
-	User UserStorage
+	User views.UserStorage
 }
